+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Optionally provide a field/method's "title" as per `@ApiModel(value = ...)`
+- Allow to ignore the general `@ApiModel(description = ...)` when populating a field/method's "description"
 
 ## [3.0.0] – 2019-06-15
 ### Added